change way of authorization

+++ microservices/star/rest/rest-core/src/main/java/pl/piomin/services/rest/customer/controller/ProductController.java
@@ -1,9 +1,12 @@
 package pl.piomin.services.rest.customer.controller;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 import pl.piomin.services.rest.customer.model.Product;
+import pl.piomin.services.rest.customer.service.AuthServiceClient;
 import pl.piomin.services.rest.customer.service.ProductServiceClient;
 
 import java.util.List;
@@ -13,25 +16,38 @@ import java.util.List;
 public class ProductController {
 
     private final ProductServiceClient client;
+    private final AuthServiceClient authClient;
 
     @GetMapping("/productAll")
     List<Product> findAll(){
-        return client.findAll();
+        if(authClient.validateUser()) {
+            return client.findAll();
+        }
+        return null;
     }
 
     @Transactional
     @PostMapping("/addProduct")
     public Product addProduct(@RequestBody Product product){
-        return client.addProduct(product);
+        if(authClient.validateUser()) {
+            return client.addProduct(product);
+        }
+        return null;
     }
 
     @PostMapping("/updateProduct")
     Product updateProduct(@RequestBody Product product){
-        return client.updateProduct(product);
+        if(authClient.validateUser()) {
+            return client.updateProduct(product);
+        }
+        return null;
     }
 
     @DeleteMapping("/deleteProduct/{name}")
     Product deleteProduct(@PathVariable String name){
-        return client.deleteProduct(name);
+        if(authClient.validateUser()) {
+            return client.deleteProduct(name);
+        }
+        return null;
     }
 }
